fix(jscad-text): move textToGeom2 to a file and fix bracket issue
